+++ src/main/kotlin/com/msg/gauth/domain/auth/presentation/AuthController.kt
@@ -3,7 +3,6 @@ package com.msg.gauth.domain.auth.presentation
 import com.msg.gauth.domain.auth.presentation.dto.request.SignUpDto
 import com.msg.gauth.domain.auth.presentation.dto.request.SigninRequestDto
 import com.msg.gauth.domain.auth.presentation.dto.response.RefreshResponseDto
-import com.msg.gauth.domain.auth.presentation.dto.response.ServiceNameResponseDto
 import com.msg.gauth.domain.auth.presentation.dto.response.SigninResponseDto
 import com.msg.gauth.domain.auth.services.*
 import org.springframework.http.HttpStatus
@@ -18,7 +17,6 @@ class AuthController(
     private val logoutService: LogoutService,
     private val signInService: SignInService,
     private val signUpService: SignUpService,
-    private val getServiceNameService: GetServiceNameService,
 ) {
     @PatchMapping
     fun refresh(@RequestHeader("RefreshToken") refreshToken: String): ResponseEntity<RefreshResponseDto> =
@@ -40,8 +38,4 @@ class AuthController(
         signUpService.execute(signUpDto)
         return ResponseEntity(HttpStatus.CREATED)
     }
-
-    @GetMapping("/{clientId}")
-    fun getServiceName(@PathVariable clientId: String): ResponseEntity<ServiceNameResponseDto> =
-        ResponseEntity.ok(getServiceNameService.execute(clientId))
 }
