--- src/main/kotlin/com/msg/gauth/domain/auth/presentation/dto/response/ServiceNameResponseDto.kt
@@ -1,5 +1,0 @@
-package com.msg.gauth.domain.auth.presentation.dto.response
-
-class ServiceNameResponseDto(
-    val serviceName: String
-)
