+++ src/main/kotlin/com/msg/gauth/domain/oauth/presentation/dto/response/ServiceNameResponseDto.kt
@@ -1,0 +1,5 @@
+package com.msg.gauth.domain.oauth.presentation.dto.response
+
+class ServiceNameResponseDto(
+    val serviceName: String
+)
