+++ src/main/kotlin/com/msg/gauth/domain/oauth/services/GetServiceNameService.kt
@@ -1,6 +1,6 @@
-package com.msg.gauth.domain.auth.services
+package com.msg.gauth.domain.oauth.services
 
-import com.msg.gauth.domain.auth.presentation.dto.response.ServiceNameResponseDto
+import com.msg.gauth.domain.oauth.presentation.dto.response.ServiceNameResponseDto
 import com.msg.gauth.domain.client.exception.ClientNotFindException
 import com.msg.gauth.domain.client.repository.ClientRepository
 import org.springframework.stereotype.Service
@@ -11,7 +11,7 @@ class GetServiceNameService(
     private val clientRepository: ClientRepository,
 ){
     @Transactional(readOnly = true, rollbackFor = [Exception::class])
-    fun execute(clientId: String): ServiceNameResponseDto{
+    fun execute(clientId: String): ServiceNameResponseDto {
         val client = (clientRepository.findByClientId(clientId)
             ?: throw ClientNotFindException())
         return ServiceNameResponseDto(
